chore: split drawer nav for better readability
