+++ packages/apps/human-app/frontend/src/router/components/layout/protected/drawer-navigation.tsx
@@ -1,22 +1,16 @@
 import Box from '@mui/material/Box';
 import Drawer from '@mui/material/Drawer';
 import CssBaseline from '@mui/material/CssBaseline';
-import List from '@mui/material/List';
-import ListItem from '@mui/material/ListItem';
-import ListItemButton from '@mui/material/ListItemButton';
-import ListItemText from '@mui/material/ListItemText';
-import { Stack, Typography } from '@mui/material';
-import { useLocation, useNavigate } from 'react-router-dom';
+import { Stack } from '@mui/material';
+import { useNavigate } from 'react-router-dom';
 import { t } from 'i18next';
 import type { Dispatch, SetStateAction } from 'react';
 import { HumanLogoNavbarIcon } from '@/shared/components/ui/icons';
 import { Button } from '@/shared/components/ui/button';
 import { useIsMobile } from '@/shared/hooks/use-is-mobile';
-import { colorPalette } from '@/shared/styles/color-palette';
-import { useColorMode } from '@/shared/contexts/color-mode';
-import { onlyDarkModeColor } from '@/shared/styles/dark-color-palette';
 import { useHandleMainNavIconClick } from '@/shared/hooks/use-handle-main-nav-icon-click';
-import { NAVBAR_PADDING } from './navbar';
+import { TopMenuItemsList } from './top-menu-items-list';
+import { BottomMenuItemsList } from './bottom-menu-items-list';
 
 const drawerWidth = 240;
 
@@ -46,11 +40,29 @@ export function DrawerNavigation({
   bottomMenuItems,
   signOut,
 }: Readonly<DrawerNavigationProps>) {
-  const { isDarkMode } = useColorMode();
   const navigate = useNavigate();
   const isMobile = useIsMobile();
-  const location = useLocation();
   const handleMainNavIconClick = useHandleMainNavIconClick();
+
+  const handleItemClick = ({ disabled, href, link, onClick }: DrawerItem) => {
+    if (onClick) {
+      onClick();
+      return;
+    }
+    if (disabled) return;
+    if (isMobile) setDrawerOpen(false);
+    if (href) {
+      const element = document.createElement('a');
+      element.href = href;
+      element.target = '_blank';
+      document.body.appendChild(element);
+      element.click();
+      return;
+    }
+    if (link && !href) {
+      navigate(link);
+    }
+  };
 
   return (
     <Box
@@ -90,188 +102,14 @@ export function DrawerNavigation({
             height: '100%',
           }}
         >
-          <List
-            sx={{
-              marginTop: isMobile ? '66px' : '16px',
-            }}
-          >
-            {topMenuItems?.map((item, index) => {
-              if (!('label' in item)) {
-                return (
-                  <ListItem key={crypto.randomUUID()}>
-                    <Stack
-                      direction="row"
-                      sx={{
-                        ml: isMobile ? '28px' : NAVBAR_PADDING,
-                      }}
-                    >
-                      {item}
-                    </Stack>
-                  </ListItem>
-                );
-              }
-
-              const { link, label, disabled, href, onClick, icon } = item;
-              const isActive = Boolean(link && location.pathname === link);
-
-              return (
-                <ListItem
-                  disablePadding
-                  key={link}
-                  sx={{ padding: '0 8px', borderRadius: '4px' }}
-                >
-                  <ListItemButton
-                    disabled={disabled}
-                    onClick={() => {
-                      if (onClick) {
-                        onClick();
-                        return;
-                      }
-                      if (disabled) return;
-                      if (isMobile) setDrawerOpen(false);
-                      if (href) {
-                        const element = document.createElement('a');
-                        element.href = href;
-                        element.target = '_blank';
-                        document.body.appendChild(element);
-                        element.click();
-                        return;
-                      }
-                      if (link && !href) {
-                        navigate(link);
-                      }
-                    }}
-                    selected={isActive}
-                    sx={{
-                      borderRadius: '4px',
-                      p: 0,
-                      '&.Mui-selected': {
-                        backgroundColor: isDarkMode
-                          ? onlyDarkModeColor.listItemColor
-                          : colorPalette.primary.shades,
-                      },
-                    }}
-                  >
-                    <Stack
-                      alignItems="center"
-                      direction="row"
-                      gap="32px"
-                      justifyContent="center"
-                      sx={{
-                        padding: '8px 16px',
-                      }}
-                    >
-                      {icon}
-                      <ListItemText
-                        disableTypography
-                        primary={
-                          <Typography
-                            component="span"
-                            variant={isMobile ? 'h5' : 'body1'}
-                          >
-                            {label}
-                          </Typography>
-                        }
-                        sx={{
-                          marginLeft: index === 0 ? '10px' : '0px',
-                        }}
-                      />
-                    </Stack>
-                  </ListItemButton>
-                </ListItem>
-              );
-            })}
-          </List>
-          <List>
-            {bottomMenuItems?.map((item) => {
-              if (!('label' in item)) {
-                return (
-                  <ListItem key={crypto.randomUUID()}>
-                    <Stack
-                      direction="row"
-                      sx={{
-                        width: '100%',
-                        mx: isMobile ? '28px' : NAVBAR_PADDING,
-                      }}
-                    >
-                      {item}
-                    </Stack>
-                  </ListItem>
-                );
-              }
-
-              const { label, link, icon, href, onClick } = item;
-              const isActive = location.pathname === link;
-
-              return (
-                <ListItem
-                  alignItems="center"
-                  disablePadding
-                  key={label}
-                  sx={{ padding: '0 8px', borderRadius: '4px' }}
-                >
-                  <ListItemButton
-                    alignItems="center"
-                    onClick={() => {
-                      if (onClick) {
-                        onClick();
-                        return;
-                      }
-                      if (isMobile) setDrawerOpen(false);
-                      if (href) {
-                        const element = document.createElement('a');
-                        element.href = href;
-                        element.target = '_blank';
-                        document.body.appendChild(element);
-                        element.click();
-                        return;
-                      }
-                      if (link && !href) {
-                        navigate(link);
-                      }
-                    }}
-                    selected={isActive}
-                    sx={{
-                      borderRadius: '4px',
-                      p: 0,
-                      '&.Mui-selected': {
-                        backgroundColor: isDarkMode
-                          ? onlyDarkModeColor.listItemColor
-                          : colorPalette.primary.shades,
-                      },
-                    }}
-                  >
-                    <Stack
-                      alignItems="center"
-                      direction="row"
-                      gap="32px"
-                      justifyContent="center"
-                      sx={{
-                        padding: '8px 16px',
-                      }}
-                    >
-                      {icon}
-                      <ListItemText
-                        disableTypography
-                        primary={
-                          <Typography
-                            component="span"
-                            variant={isMobile ? 'h5' : 'body1'}
-                          >
-                            {label}
-                          </Typography>
-                        }
-                        sx={{
-                          textAlign: 'center',
-                          marginLeft: '10px',
-                        }}
-                      />
-                    </Stack>
-                  </ListItemButton>
-                </ListItem>
-              );
-            })}
-          </List>
+          <TopMenuItemsList
+            handleItemClick={handleItemClick}
+            items={topMenuItems}
+          />
+          <BottomMenuItemsList
+            handleItemClick={handleItemClick}
+            items={bottomMenuItems}
+          />
         </Stack>
         <Button
           onClick={() => {
